Add files via upload
In the above code, we first perform residual analysis by calculating the residuals (i.e., the differences between the actual and predicted target variables) for the training set. We then check the distribution of error terms using the display function from the Seaborn library, which creates a histogram of the residuals.